According to the spec, when cloning a typed array that is a view of a larger buffer, the entire buffer should be cloned and then the view recreated on that buffer. Firefox used to copy only the portion within the view, which observably differs from the spec. Its behavior is changing, which can result in hugely inflated memory usage.

+++ Player/mp4.js
@@ -866,7 +866,9 @@ var MP4Player = (function reader() {
         if (this.useWorkers) {
           var avcWorker = this.avcWorker;
           video.getSampleNALUnits(pic).forEach(function (nal) {
-            avcWorker.sendMessage("decode-sample", nal);
+            // Copy the sample so that we only do a structured clone of the
+            // region of interest
+            avcWorker.sendMessage("decode-sample", Uint8Array(nal));
           });
         } else {
           var avc = this.avc;
